renamed RedisPubsub, fixed helm condition
