+++ generators/app/templates/infrastructure/helm/gql/templates/deployment.yaml
@@ -81,7 +81,7 @@ spec:
           - name: DB_DATABASE
             value: {{ $global.db.database | quote }}
 <%_ if(addVaultConfigs){ _%>
-          {{- if not $global.vaultEnvironment | eq "false" }}
+          {{- if $global.vaultEnvironment | eq "false" }}
 <%_}_%>
           - name: DB_USER
             value: {{ $global.db.user | quote }}
