+++ generators/app/templates/infrastructure/src/pubSub/index.js
@@ -1,2 +1,2 @@
 module.exports.topics = require("./topics")
-module.exports.pubSub = require("./redisPubSub")
+module.exports.pubSub = require("./pubSub")
