+++ generators/app/templates/infrastructure/src/pubSub/pubSub.js
